# openapi/pulse-api.yaml
openapi: 3.0.0
paths:
...
  /tracks/{id}/stream:
    get:
      description: Devuelve el MP3 entero, o el trozo pedido si llega una cabecera `Range`. El reproductor necesita rangos para poder saltar por la canción sin descargarla completa.
      operationId: streamTrack
      parameters:
        - name: id
          required: true
          in: path
          description: UUID de la pista, su slug canónico, o un identificador antiguo `tr-*`.
          schema:
            example: get-lucky
            type: string
        - name: Range
          in: header
          description: Por ejemplo `bytes=0-65535`.
          required: false
          schema:
            type: string
        - name: sig
          required: true
          in: query
          description: Firma que emite el catálogo junto a la URL.
          schema: {}
        - name: exp
          required: true
          in: query
          description: Caducidad de la firma, en segundos desde la época.
          schema: {}
        - name: v
          required: false
          in: query
          description: Marca de versión que emite el catálogo. Solo rompe la caché del cliente cuando se reemplaza el audio; el servidor la ignora.
          schema: {}
      responses:
        "200":
          description: MP3 completo.
        "206":
          description: Trozo del MP3 pedido por `Range`.
        "403":
          description: La firma falta, no cuadra, o ya caducó. Vuelve a pedir `/catalog`.
        "404":
          description: No existe ninguna pista con ese identificador.
        "416":
          description: El `Range` pedido cae fuera del archivo.
      summary: Reproducir una pista
      tags:
        - Reproducción
...
info:
  title: Pulse Music API
  description: |-
    Catálogo y reproducción de la biblioteca de Pulse.

    La autenticación es global: salvo lo marcado como público —`/health`, `/catalog`,
    `/tracks/{id}/stream`, `/media/covers/{file}` y el propio `/auth`—, toda ruta exige sesión.

    `/tracks/{id}/stream` es público en el sentido de que no pide un Bearer: un `<audio src>` no
    puede mandar cabeceras. En su lugar exige la firma que emite `/catalog` en cada `streamUrl`,
    válida 24 horas. Sin ella responde 403, y la forma de conseguir una nueva es volver a pedir
    el catálogo.

    Lo que hay bajo `/admin` mantiene el catálogo y pide rol `ADMIN`. Se entra de dos maneras:
    con la sesión de una cuenta administradora, o con la cabecera `X-Pulse-Admin-Token`, que
    queda reservada a los procesos automatizados. Un credencial válido pero sin rol recibe 403,
    no 401: no hay nada que renovar.

    Lo que hay bajo `/auth` abre y renueva sesiones. El access token es un JWT de vida corta que
    viaja en `Authorization: Bearer`; el refresh token dura semanas y sale en una cookie
    `httpOnly` que solo se manda a `/v1/auth`. Un cliente sin cookies —la app móvil— lo pide en
    el cuerpo con la cabecera `X-Pulse-Client: native`.

    Toda ruta cuelga del prefijo `/v1`, ya incluido en la URL del servidor.
  version: 0.1.0
  contact: {}
...
servers:
  - url: http://localhost:4000/v1
    description: Servidor
components:
  securitySchemes:
    AdminToken:
      type: apiKey
      in: header
      name: X-Pulse-Admin-Token
      description: "Credencial de servicio para procesos sin persona detrás, igual a `PULSE_ADMIN_TOKEN` en el servidor. No es para el navegador: desde la web se entra con una cuenta de rol ADMIN."
    BearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: "Access token devuelto por `/auth/login`. Caduca a los 15 minutos: renuévalo en `/auth/refresh`."
...